+++ src/components/ui/Dashboard_ui/TableUserOverview.js
@@ -2,7 +2,7 @@ import {Link, useHistory} from "react-router-dom";
 import {useEffect, useState} from "react";
 import PropTypes from "prop-types";
 import {api, handleError, get_with_token, api_with_token} from "../../../helpers/api";
-import UserList from "./UserList";
+import UserList from "../UserList";
 import {Button} from "../Button";
 import BaseContainer from "../BaseContainer";
 import "styles/_theme.scss";
